+++ doc/CHANGELOG.md
@@ -1,3 +1,8 @@
+[0.2.4]
+* Fix bug in bin/healthcheck.py to cast VEIN_QUERY_PORT and VEIN_GAME_PORT to int
+* Fixed chmod path in backup.Dockerfile
+* update backup script to use trap for graceful shutdown
+
 [0.2.3]
 
 * Added support to use S3 to backup Game.ini and Server.vns